Clean up env options (#498)
- Move enable options to options.h/cpp
- Add a options guard to temporarily change option values.

Currently only the enable options are modified. Will update the rest.

Update: moved DebugDump and Disable options too
