+++ csrc/codegen.cpp
@@ -11,6 +11,7 @@
 #include <ir/utils.h>
 #include <kernel_ir.h>
 #include <kernel_ir_dispatch.h>
+#include <options.h>
 #include <scheduler/mma_utils.h>
 #include <type.h>
 #include <utils.h>
