+++ csrc/device_lower/pass/expr_sort.cpp
@@ -14,6 +14,7 @@
 #include <ir/all_nodes.h>
 #include <ir/iostream.h>
 #include <ir/utils.h>
+#include <options.h>
 #include <utils.h>
 
 #include <deque>
