+++ csrc/expr_simplifier.cpp
@@ -13,6 +13,7 @@
 #include <ir/cloner.h>
 #include <ir/iostream.h>
 #include <ir/utils.h>
+#include <options.h>
 #include <utils.h>
 
 #include <cmath>
